requirements.txt
flask==3.1.0
requests==2.32.3
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.20.0
google-cloud-secret-manager==2.21.0
vertexai==1.71.0
gunicorn==23.0.0
